requirements.txt
aiohttp~=3.8.4
bs4
lxml
pymysql
python-dotenv
aiomysql